Created Domain + Repository pattern
First of all sorry for this commit, but I needed to create a base architecture in order to understand well how it works.

I have re-structured MVP pattern in instagrammers package in a more understandable way (view contains activities, holders and adapters).
I have created a view model and a domain model. Also did the interactor, a repository and a datasource (it returns fake data for the moment).
Presenter have been connected to interact and I've created mappers.