Feature/search instagrammers
* Created new activity

Created activity to do the intagrammers search

* Basic MVP

It has been done all the basic structure (activity, presenter and interactor that returns an empty list when someone searchs). It has been necessary to update Rosie to v2 in order to use RosieAppCompatActivity, so we can do the search view.

* Done Repository calls

Repository calls have been implemented as well as three new Objects in order to obtain the response. It's been used Retrofit 1.9.

* Added permission when authenticating in the app

This permission is necessary to search another instagrammers

* Datasource Refactored

Extracted a method used to create the service

* New Adapter and holder to show instagrammers searched

It has been necessary to create another mapper for searched instagrammers

* Added error handling to the view

* Presenter covered with tests

* Use cases renamed

It's easier to use and find them if we add the postscript "UseCase"

* Renamed Repositories and Datasources

Its responsibilities are not with Followed Instagrammers only, but with all operations related with instagrammers

* Created a service generator

So there is no need to duplicate code each time you do a call to and endpoint.It's injected in the datasource.

* Rebased onto develop